seedファイルの作成、DB構造の修正、テストの追加

+++ db/migrate/20190325084924_create_skills.rb
@@ -1,8 +1,8 @@
 class CreateSkills < ActiveRecord::Migration[5.2]
   def change
     create_table :skills do |t|
-      t.integer :skill_type_id
-      t.integer :skill_detail_type_id
+      t.bigint :skill_type_id
+      t.bigint :skill_detail_type_id
       t.string :name
       t.string :code
